# tools/sdk-ts/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Notifo API",
    "version": "1.0.0",
    "x-logo": {
      "url": "/logo.svg"
    }
  },
  "servers": [
    {
      "url": "https://localhost:5002"
    }
  ],
  "paths": {
...
    "/api/apps": {
      "get": {
        "tags": [
          "Apps"
        ],
        "summary": "Get the user apps.",
        "operationId": "Apps_GetApps",
        "responses": {
          "200": {
            "description": "Apps returned.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "getApps"
      },
      "post": {
        "tags": [
          "Apps"
        ],
        "summary": "Create an app.",
        "operationId": "Apps_PostApp",
        "requestBody": {
          "x-name": "request",
          "description": "The request object.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertAppDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "App created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "postApp"
      }
    },
    "/api/apps/{appId}": {
      "get": {
        "tags": [
          "Apps"
        ],
        "summary": "Get app by ID.",
        "operationId": "Apps_GetApp",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "description": "The ID of the app.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "App returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "App not found."
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "getApp"
      },
      "post": {
        "tags": [
          "Apps"
        ],
        "summary": "Update an app.",
        "operationId": "Apps_PutApp",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "description": "The app id to update.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "description": "The request object.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertAppDto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "App updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "App not found."
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "putApp"
      }
    },
    "/api/apps/{appId}/auth": {
      "get": {
        "tags": [
          "Apps"
        ],
        "summary": "Get app auth settings by ID.",
        "operationId": "Apps_GetAuthScheme",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "description": "The ID of the app.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "App auth settings returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSchemeValueDto"
                }
              }
            }
          },
          "404": {
            "description": "App not found."
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "getAuthScheme"
      },
      "put": {
        "tags": [
          "Apps"
        ],
        "summary": "Updates the auth settings of the app.",
        "operationId": "Apps_UpsertAuthScheme",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "description": "The ID of the app.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "description": "The request object.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthSchemeValueDto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "App auth settings returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSchemeValueDto"
                }
              }
            }
          },
          "404": {
            "description": "App not found."
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "upsertAuthScheme"
      }
    },
    "/api/apps/{appId}/contributors": {
      "post": {
        "tags": [
          "Apps"
        ],
        "summary": "Add an app contributor.",
        "operationId": "Apps_PostContributor",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "description": "The ID of the app.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "description": "The request object.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddContributorDto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "Apps returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "App not found."
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "postContributor"
      }
    },
    "/api/apps/{appId}/contributors/{contributorId}": {
      "post": {
        "tags": [
          "Apps"
        ],
        "summary": "Delete an app contributor.",
        "operationId": "Apps_DeleteContributor",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "description": "The ID of the app.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "contributorId",
            "in": "path",
            "required": true,
            "description": "The contributor to remove.",
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "Apps returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "App not found."
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "deleteContributor"
      }
    },
    "/api/apps/{appId}/integrations": {
      "get": {
        "tags": [
          "Apps"
        ],
        "summary": "Get the app integrations.",
        "operationId": "Apps_GetIntegrations",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "description": "The ID of the app where the integrations belong to.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "App email templates returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfiguredIntegrationsDto"
                }
              }
            }
          },
          "404": {
            "description": "App not found."
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "getIntegrations"
      }
    },
    "/api/apps/{appId}/integration": {
      "post": {
        "tags": [
          "Apps"
        ],
        "summary": "Create an app integrations.",
        "operationId": "Apps_PostIntegration",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "description": "The ID of the app where the integration belong to.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "description": "The request object.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIntegrationDto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "App integration created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationCreatedDto"
                }
              }
            }
          },
          "404": {
            "description": "App not found."
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "postIntegration"
      }
    },
    "/api/apps/{appId}/integrations/{id}": {
      "put": {
        "tags": [
          "Apps"
        ],
        "summary": "Update an app integration.",
        "operationId": "Apps_PutIntegration",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "description": "The ID of the app where the integration belong to.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the integration.",
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "request",
          "description": "The request object.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIntegrationDto"
              }
            }
          },
          "required": true,
          "x-position": 3
        },
        "responses": {
          "204": {
            "description": "App integration updated."
          },
          "404": {
            "description": "App not found."
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "putIntegration"
      },
      "delete": {
        "tags": [
          "Apps"
        ],
        "summary": "Delete an app integration.",
        "operationId": "Apps_DeleteIntegration",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "description": "The ID of the app where the email templates belong to.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the integration.",
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "204": {
            "description": "App integration deleted."
          },
          "404": {
            "description": "App not found."
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "deleteIntegration"
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "statusCode"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "The error message.",
            "nullable": true
          },
          "errorCode": {
            "type": "string",
            "description": "The error code.",
            "nullable": true
          },
          "traceId": {
            "type": "string",
            "description": "The optional trace id.",
            "nullable": true
          },
          "type": {
            "type": "string",
            "description": "The error type, usually a link.",
            "nullable": true
          },
          "details": {
            "type": "array",
            "description": "Detailed error messages.",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "statusCode": {
            "type": "integer",
            "description": "Status code of the http response.",
            "format": "int32"
          }
        }
      },
...
      "AppDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "name",
          "role",
          "created",
          "lastUpdate",
          "languages",
          "apiKeys",
          "counters"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The ID of the app."
          },
          "name": {
            "type": "string",
            "description": "The app name."
          },
          "role": {
            "type": "string",
            "description": "The current role."
          },
          "created": {
            "type": "string",
            "description": "The date time (ISO 8601) when the app has been created.",
            "format": "date-time"
          },
          "lastUpdate": {
            "type": "string",
            "description": "The date time (ISO 8601) when the app has been updated.",
            "format": "date-time"
          },
          "languages": {
            "type": "array",
            "description": "The supported languages.",
            "items": {
              "type": "string"
            }
          },
          "apiKeys": {
            "type": "object",
            "description": "The api keys.",
            "additionalProperties": {
              "type": "string"
            }
          },
          "counters": {
            "type": "object",
            "description": "The statistics counters.",
            "additionalProperties": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
      "AppDetailsDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "name",
          "role",
          "created",
          "lastUpdate",
          "languages",
          "apiKeys",
          "contributors",
          "counters"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The ID of the app."
          },
          "name": {
            "type": "string",
            "description": "The app name."
          },
          "role": {
            "type": "string",
            "description": "The current role."
          },
          "created": {
            "type": "string",
            "description": "The date time (ISO 8601) when the app has been created.",
            "format": "date-time"
          },
          "lastUpdate": {
            "type": "string",
            "description": "The date time (ISO 8601) when the app has been updated.",
            "format": "date-time"
          },
          "confirmUrl": {
            "type": "string",
            "description": "The confirm URL.",
            "nullable": true
          },
          "languages": {
            "type": "array",
            "description": "The supported languages.",
            "items": {
              "type": "string"
            }
          },
          "apiKeys": {
            "type": "object",
            "description": "The api keys.",
            "additionalProperties": {
              "type": "string"
            }
          },
          "contributors": {
            "type": "array",
            "description": "The contributors.",
            "items": {
              "$ref": "#/components/schemas/AppContributorDto"
            }
          },
          "counters": {
            "type": "object",
            "description": "The statistics counters.",
            "additionalProperties": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
...
      "AuthSchemeValueDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "scheme": {
            "description": "The auth scheme if configured.",
            "nullable": true,
            "$ref": "#/components/schemas/AuthSchemeDto"
          }
        }
      },
...
      "UpsertAppDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string",
            "description": "The app name.",
            "nullable": true
          },
          "languages": {
            "type": "array",
            "description": "The supported languages.",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "emailAddress": {
            "type": "string",
            "description": "The sender email address.",
            "nullable": true
          },
          "emailName": {
            "type": "string",
            "description": "The sender email name.",
            "nullable": true
          },
          "firebaseProject": {
            "type": "string",
            "description": "The firebase project ID.",
            "nullable": true
          },
          "firebaseCredential": {
            "type": "string",
            "description": "The firebase credentials.",
            "nullable": true
          },
          "webhookUrl": {
            "type": "string",
            "description": "The webhook URL.",
            "nullable": true
          },
          "confirmUrl": {
            "type": "string",
            "description": "The confirm URL.",
            "nullable": true
          },
          "allowEmail": {
            "type": "boolean",
            "description": "True, when emails are allowed.",
            "nullable": true
          },
          "allowSms": {
            "type": "boolean",
            "description": "True, when SMS are allowed.",
            "nullable": true
          }
        }
      },
      "AddContributorDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string",
            "description": "The email of the new contributor.",
            "minLength": 1
          },
          "role": {
            "type": "string",
            "description": "The role.",
            "nullable": true
          }
        }
      },
...
    }
  }
}